hack-yolo-worker/pyproject.toml
[project]
name = "server"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.34.69",
    "botocore>=1.34.69",
    "kafka-python>=2.2.6",
    "ultralytics>=8.3.131",
    "fastapi>=0.115.12",
    "uvicorn>=0.34.2",
    "ultralytics>=8.3.145",
    "standard-imghdr>=3.13.0",
    "httpx>=0.28.1",
]
